refactor: clarify app and explorer ownership
