+++ src/pages/explorer/widgets/ExplorerSidebar/ui/ExplorerSidebar.tsx
@@ -2,10 +2,10 @@ import { useAtom, useAtomValue } from 'jotai';
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import type React from 'react';
 import { useRef, useState } from 'react';
+import { fileTreeModeAtom, isSidebarOpenAtom } from '@/features/Explorer/model/atoms';
 import { Sidebar } from '@/shared/ui/Sidebar';
 import { FileExplorer } from '@/widgets/FileExplorer/FileExplorer.tsx';
 import { RelatedFilesView } from '../../RelatedFilesView/ui/RelatedFilesView';
-import { fileTreeModeAtom, isSidebarOpenAtom } from '../model/atoms.ts';
 
 export const ExplorerSidebar: React.FC = () => {
   const isSidebarOpen = useAtomValue(isSidebarOpenAtom);
