+++ src/widgets/JotaiDevTools/JotaiDevTools.tsx
@@ -12,10 +12,10 @@ import * as canvasAtoms from '@/features/Canvas/model/atoms';
 import * as deadCodeAtoms from '@/features/Code/CodeAnalyzer/DeadCodeAnalyzer/model/atoms';
 import * as codeFoldAtoms from '@/features/Code/CodeFold/model/atoms';
 import * as focusModeAtoms from '@/features/Code/FocusMode/model/atoms';
+import * as explorerAtoms from '@/features/Explorer/model/atoms';
 import * as navigationAtoms from '@/features/File/Navigation/model/atoms';
 import * as filesAtoms from '@/features/File/OpenFiles/model/atoms';
 import * as searchAtoms from '@/features/Search/UnifiedSearch/model/atoms';
-import * as sidebarAtoms from '@/pages/explorer/widgets/ExplorerSidebar/model/atoms';
 
 // Combine all atoms for DevTools tracking
 const atoms = {
@@ -28,7 +28,7 @@ const atoms = {
   ...deadCodeAtoms,
   ...filesAtoms,
   ...canvasAtoms,
-  ...sidebarAtoms,
+  ...explorerAtoms,
 };
 
 interface AtomUpdate {
